Move html test to HighlightedStringTests file
