+++ Tests/AlgoliaSearchClientTests/Unit/HighlightedStringTests.swift
@@ -55,6 +55,17 @@ class HighlightedStringTests: XCTestCase {
     
   }
 
+    struct MarkupString: Codable {
+      let source: String
+      let markup: HighlightedString
+    }
+  
+    func testWithHTMLString() throws {
+      let data = try Data(filename: "HighlightedHTML.json")
+      let searchResponse = try JSONDecoder().decode(MarkupString.self, from: data)
+      let searchResponseJSON = try JSON(searchResponse)
+      print(searchResponseJSON)
+    }
 
 }
